# openapi.yaml
openapi: 3.1.0
info:
  title: Patient Navigator API
  version: 0.1.0
  description: Content and neutral reminder API for the MAX patient navigator MVP.
servers:
  - url: http://localhost:3001
    description: Local development API
paths:
  /health:
    get:
      summary: Liveness check
      responses:
        "200":
          description: API process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /ready:
    get:
      summary: Readiness check
      description: Confirms that the API can load and validate the content catalog.
      responses:
        "200":
          description: API is ready to serve traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ready"
        "503":
          description: API is alive but content is unavailable
  /content/version:
    get:
      summary: Content catalog version
      responses:
        "200":
          description: Current content metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentVersion"
  /situations:
    get:
      summary: List situation cards
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Optional case-insensitive substring search across title, keywords and steps.
      responses:
        "200":
          description: Situation cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SituationCard"
  /situations/{id}:
    get:
      summary: Get one situation card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Situation card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SituationCard"
        "404":
          description: Situation not found
  /knowledge:
    get:
      summary: List knowledge cards
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Optional case-insensitive substring search.
      responses:
        "200":
          description: Knowledge cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeCard"
  /reminders:
    post:
      summary: Create or update a neutral reminder
      description: Reminder text must be neutral and must not contain diagnosis or medical details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "200":
          description: Existing reminder updated idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "400":
          description: Validation error
  /reminders/{id}:
    get:
      summary: Get reminder by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Reminder not found
    delete:
      summary: Cancel reminder by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reminder cancelled
        "404":
          description: Reminder not found
components:
  schemas:
    Health:
      type: object
      required: [status, service]
      properties:
        status:
          const: ok
        service:
          const: navigator-api
    Ready:
      type: object
      required: [status, service, content]
      properties:
        status:
          const: ready
        service:
          const: navigator-api
        content:
          $ref: "#/components/schemas/ContentVersion"
    ContentVersion:
      type: object
      required: [version, situations, knowledge, dataStatus]
      properties:
        version:
          type: string
        situations:
          type: integer
          minimum: 0
        knowledge:
          type: integer
          minimum: 0
        dataStatus:
          type: string
        commit:
          type: string
    SourceRef:
      type: object
      required: [title, type, actualOn]
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
        type:
          enum: [official, clinic, law, faq, synthetic]
        actualOn:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
    DecisionQuestion:
      type: object
      required: [id, text, type, options]
      properties:
        id:
          type: string
        text:
          type: string
        type:
          const: single
        options:
          type: array
          minItems: 2
          items:
            type: object
            required: [value, label]
            properties:
              value:
                type: string
              label:
                type: string
              note:
                type: string
    SituationStep:
      type: object
      required: [id, title, summary, why, whereToGet, commonMistakes, sourceIds]
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        why:
          type: string
        whereToGet:
          type: array
          minItems: 1
          items:
            type: string
        validFor:
          type: string
        commonMistakes:
          type: array
          items:
            type: string
        sourceIds:
          type: array
          minItems: 1
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        conditions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    SituationCard:
      type: object
      required: [id, title, category, shortDescription, region, dataStatus, actualOn, reviewBefore, disclaimer, keywords, decisionQuestions, steps, knowledgeLinks, sources]
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        region:
          type: string
        dataStatus:
          enum: [official, clinic, synthetic]
        actualOn:
          type: string
        reviewBefore:
          type: string
        disclaimer:
          type: string
        keywords:
          type: array
          items:
            type: string
        decisionQuestions:
          type: array
          items:
            $ref: "#/components/schemas/DecisionQuestion"
        steps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SituationStep"
        knowledgeLinks:
          type: array
          items:
            type: string
        sources:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SourceRef"
    KnowledgeCard:
      type: object
      required: [id, title, summary, body, tags, sourceIds, sources, actualOn, dataStatus]
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        body:
          type: array
          minItems: 1
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        sourceIds:
          type: array
          items:
            type: string
        sources:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SourceRef"
        actualOn:
          type: string
        dataStatus:
          enum: [official, clinic, synthetic]
    ReminderRequest:
      type: object
      required: [situationId, stepId, remindAt, text]
      properties:
        chatId:
          type: string
        situationId:
          type: string
        stepId:
          type: string
        remindAt:
          type: string
          format: date-time
        text:
          type: string
          minLength: 1
          maxLength: 240
    Reminder:
      allOf:
        - $ref: "#/components/schemas/ReminderRequest"
        - type: object
          required: [id, status, createdAt, updatedAt]
          properties:
            id:
              type: string
            status:
              enum: [pending, sent, cancelled]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time